Delete `username` from `users` table and add `id` to `UserPrincipal` as user account information
There isn't a specific need to use the `username` for the purposes of the project, and since the `email` field also has to be UNIQUE, the `username` is not necessary.

`Because of this change, a project-specific version of `UserDetails` and `UserDetailsService` was implemented, since it doesn't suit the project needs.