Add `userExists(Long id)`, `getUserById`, and `loadUserPrincipalById` to `UserService`.
Refactor `UserServiceImpl` to retrieve users using `getUserById` by default.

Remove `loadPrincipalByEmail` and `getUserByPrincipal` from `UserService`.

Update `withUser` of `UserPrincipalImpl` to `fromUser` to indicate that it builds a `UserPrincipalImpl` with exact information from `User`.

Add `withUser` method to `UserPrincipalImpl` to indicate that it uses information from `User`, but it can be modified with the returned `UserPrincipalImplBuilder` before building the `UserPrincipalImpl` object.

Update `JwtService` to use the user's ID as subject rather than its email.